Restructure abilities registration to an ability-per-class basis
